fixing the standard header pointers for it to work

+++ slsReceiverSoftware/src/DataProcessor.cpp
@@ -294,7 +294,7 @@ void DataProcessor::StopProcessing(char* buf) {
 #endif
 }
 
-
+/** buf includes only the standard header */
 void DataProcessor::ProcessAnImage(char* buf) {
 	numFramesCaught++;
 	numTotalFramesCaught++;
@@ -302,12 +302,12 @@ void DataProcessor::ProcessAnImage(char* buf) {
 	sls_detector_header* header = (sls_detector_header*) (buf);
 	uint64_t fnum = header->frameNumber;
 #ifdef VERBOSE
-	if (!index) cprintf(BLUE,"DataProcessing %d: fnum:%lld\n", index, (long long int)fnum);
+	if (!index) cprintf(BLUE,"DataProcessing %d: fnum:%lu\n", index, fnum);
 #endif
 
 	if (!measurementStartedFlag) {
 #ifdef VERBOSE
-		if (!index) cprintf(BLUE,"DataProcessing %d: fnum:%lld\n", index, (long long int)fnum);
+		if (!index) cprintf(BLUE,"DataProcessing %d: fnum:%lu\n", index, fnum);
 #endif
 		RecordFirstIndices(fnum);
 	}
